feat: adiciona validação verificando se o produto existe

+++ src/main/java/com/example/retornosAPI/services/ProductService.java
@@ -39,6 +39,9 @@ public class ProductService {
     }
 
     public void deleteProduct(Long id) {
+        if (!repository.existsById(id)) {
+            throw new RuntimeException("Produto não encontrado.");
+        }
         repository.deleteById(id);
     }
